Patches #4, Newtonsoft.Json.JsonReaderException : Input string '7999.0' is not a valid integer. Path 'button.price.cents', line 1, position 660.

+++ Source/Coinbase/Order.cs
@@ -107,7 +107,7 @@ namespace Coinbase
 
     public class Price
     {
-        public int Cents { get; set; }
+        public decimal Cents { get; set; }
 
         [JsonProperty("currency_iso")]
         public Currency Currency { get; set; }
